fix(server): model state fees as uint64 drops

+++ xrpl/rpc/helpers.go
@@ -435,7 +435,7 @@ func (c *Client) fetchOwnerReserveFee(ctx context.Context) (uint64, error) {
 		return 0, ErrCouldNotFetchOwnerReserve
 	}
 
-	return uint64(*reserveInc), nil
+	return *reserveInc, nil
 }
 
 // fetchCounterPartySignersCount fetches the number of signers for the counterparty account.
